Consistency changes to Sim class.
Removed the word 'base' from use (mostly), as it is annoyingly
ambiguous. Also, added proper exception for no projectdir being
provided.

Changed arrangement for automatic object storage. Now objects are stored
in a directory called MDSynthesis in the projectdir, and they are
arranged according to the type of container. So, Sim objects will be
placed in `*projectdir*/MDSynthesis/Sim`, SimSet objects in
`*projectdir*/MDSynthesis/SimSet`, etc. This allows for objects
unforeseen to be added to the package seamlessly.