Many changes to Sim with regard to path metadata.
Sim will no longer take a universe as an input. A Sim has two ways of being
composed: it can be generated for the first time using the same input arguments
and keywords as a Universe; it can be re-generated by providing a base
directory where a metadata file (__name__.yaml) resides. This simplifies the
usage and API greatly.

Also, all paths in the metadata directory are stored with respect to the
new 'project' directory. This is the directory in which, presumably, one has
all trajectories associated with a particular project. After much deliberation,
this seemed like an appropriate scheme, with the fewest drawbacks. One can
specify the projectdir to be anything one wants, so it actually does provide
some flexibility without sacrificing the core strengths of the scheme. One can
still specify a location keyword to stick the object wherever is desired. The
behavior of this keyword may change, as it might give a result that is overly
complicated by most standards.

The projectdir and basedir now should change dynamically if they change in the
filesystem. This means that if one moves the projectdir to a new place in an
old or new filesystem, objects will still work (so long as trajectory paths
relative to this projectdir don't ALSO change). Also, one can now move objects
around the MDSynthesis directory however one likes (even outside of it, so long
as they are still in the projectdir), and they will update their location on
regeneration.

Also, finally, many new under-the-hood methods were added to make all of this
new infrastructure work. Mainly the changes were to add better organization.

Oh, and logging is in the same format as MDAnalysis. :D