Finished first iteration of Group.
Uses eval() to reload individual Sim-derived containers based on
the type of container they correspond to. This is going to be very
sensitive to how these classes exist in the namespace, which is a
downside to this method.

The upshot is that you can, in principle, have a Group object that
supports having ANY combination of Sim-derived objects as members.

It might be better, however, to make the Group class specific to Sim
objects, building in mechanisms for inheritance that allow it to work
on user-defined Sim-derived objects with specific adaptation.