Finished Group container.
As written, it only works for Sim objects, but it is written in such
a way to be expanded easily to any number of Sim-derived objects. In fact,
I have already done this for three such objects in one of my own projects.
They are all handled by one Group-derived object.