Added the ability to write out data files to Containers.
Now a container can write out loaded data instances to their source files. This
might be useful in those cases where one might need to manually add to or edit
the data files (VERY RARE), and so this provides a convenient and safe way to
do it. One must explicitly call the save() method with specific arguments to
make the save happen.