Re-working containers for use with database.
Now that the new database scheme has been (mostly) hashed out, I am in the
process of implementing the details in the ContainerCore and Sim (for a start).
I am doing my best to implement as much as I can in ContainerCore, such that
all Containers operate in a predictable and uniform way under this new scheme.

I am still working out the internal structure of the database. I imagine it
will be a dictionary with 'metadata' and 'containers' top-level keys. The
metadata key will itself have a dictionary value, with information relevant to
the database (its location, id, name(?) etc.). The database having an id or a
name is something I hadn't thought of until now, but it might prove useful for
containers to know not only where their database ought to live, but what its id
is as well. This would allow for the following scheme in case a Container can't
find its database:

1) searches upward for database, checking id of each it finds until it finds
ITS database.
2) if it didn't find a database that matches the id on file, it takes the first
it finds in the upward search it did
3) if no database was found in the upward search, a new database is built where
the old one should have been; if the Container never had a database (such as
when it is just being generated), an exception is thrown.

This scheme would allow one to have objects associated with a certain database
to exist in the "realm" of another. This might get confusing, actually, but it
might also allow for some unforeseen flexibility. I'll have to chew on this for
a little bit. An even crazier idea is to have things built in such a way that
a Container can associate with multiple databases (just like a Sim can have
multiple universes), which would probably result in more headaches to ensure it
all works that it would be worth.

Perhaps, for now just keep things single-database, with the ability for
multi-database functionality later without breaking the API.