Hammering away at Database.
Search string functionality will have to be added later. In the near term,
might add simple search to match criteria using AND mode or OR mode.
This is hardly elegant, but it would be useful until a more thorough solution
is in place.

Almost finished with Database basic functionality. Need to finish up container
push and pull (might even rename these to those terms). Not certain if I'll
include options to only push or pull specified keys, but I might just try to
see if it can be done simply enough. It doesn't seem like a really necessary
feature, but who knows...

Still need to figure out some kind of file-lock scheme, but this can come
later. It will only become an issue if many Containers are writing to the
database at once. This might become a real challenge.