Discover now works; can easily add all containers below db.
This is damn good news. The next item on the agenda is to fix the Group.
Here are thoughts.

For the sake of consistency, I would love to make the 'member' key in
the Group metadata refer to a dictionary, with each entry keyed with a
uuid and with nothing else but a path and class type. However, a list
is kind of nice in its own way, despite the fact we don't care about
order here.

It's probably better to go with the dictionary, for three reasons:
    1) uuid as key, though slightly cumbersome for interactive use, will
    always get you the object you expect it to. Always. No need to
    depend on order of entries in metadata.
    2) order isn't a need, which is one of the main selling point of a
    list over a dict.
    3) adding to the set of containers is easy, since indexing by uuid
    ensures no doubles. No need to put in machinery to remove doubles.

It's too difficult to write a converter to fix existing Group objects.
I will have to simply remake them manually. If I make copies, I'll avoid
catastrophe.