Finished Sim init, and low-level database interfaces.
These are all untested, and push and pull have yet to be implemented.
However, it is all looking very good. Tomorrow will have to finish
refactoring Group, as that got pushed to the side while I was working on
this more general problem.

Once Group is refactored, we then refactor the Operators. Hopefully
those won't require too much work...