File locking implemented.
Now, all objects have a utilities attribute with an 'open' method. This
method can be used just like the 'open' builtin, but it will perform
automatic exclusive file locking on the file being opened. This method
should be used every time an object needs to open a file, as it ensure
race conditions cannot occur.

Now that file locking is taken care of under the hood, we can now do
file i/o without worrying about race conditions on write. Typical usage
of utilities.open, however, will not ensure that old data in a loaded
object is not written to a since-updated file. To ensure this, it might
be worthwhile to build a standard cache scheme like this:

object._metadata -> cached metadata; always reflect last read of file
object.metadata -> copy of cached metadata

object.save() will first compare cached metadata to its userland
copy (which may have been modified by hand, by normal action of object,
etc.). Differences will be recorded...ah wait, this won't work. Too
confusing to build an algorithm with behavior one would expect without
it having the ability to read minds.

We may have to make a metadata object, and build methods to modify its
contents. With atomically applied operations, this will ensure that the
version in the object, when changed, always syncs these changes properly
to the file on disk. We should also build a database (lowercase) object
for database files, though these might not need to be separate.

Food for thought. Things are coming together nicely for this.
Reliable parallelism with persistence is looking better and better.