File locking no longer uses fcntl; more robust for NFS.
File locking in POSIX is a mess, and there is no guarantee that on a networked
filesystem file-locking as currently implemented will work across these
systems. As such, I re-implemented the file-locking scheme to use a more
traditional lockfile approach.

Unfortunately, this is rather slow. If there are ways around this, we will have
to explore them, but it ultimately comes down to filesystem i/o.

Not entirely sure if this guarantees atomistic transactions on NFS, but it's a
better bet than the old scheme.