Adding method for Database to get containers on demand.
When a Database must push information out to Containers, it needs to verify
that it is sending it to the right recipient. The _get_container method will
verify that the Container being addressed is the correct one based on UUID, and
if it isn't (or if it can't be found at all), it will look downward through the
filesystem until it is found. If it isn't found, an exception will occur, or at
least a warning will be logged.