Database can now locate Containers when they go missing.
The Database interacts with Containers in two fundamental ways:
pushing metadata to them, and pulling metadata from them. When a push or
a pull is attempted, but the Container is not found where it is
expected, the Database goes looking for it. If it is not found a warning
is given, and the object is not updated.

Unlike the Container functionality for finding a Database once it has
gone missing, the Database ensures it finds the object it is looking for
based on uuid. To contrast, a Container that goes looking for its
Database attaches to the first one it finds, even if it is different
than its previous one.