Added cache, decache to Sim.
Untested feature: must test with non-production trajectory before use!

Sim object can now make a copy of a universe's structure and trajectory
elsewhere and use that as the basis for that universe. This allows, for
example, the copying of a large trajectory from a remote filesystem to local
scratch space before performing analysis. To clean up, the user need only use
the decache method, specifying the universe to cache.