Cache will now abort if another instance detected.
When attempting to cache, if another instance of a container is cacheing
the same universe in the same location, the container will abort the
cache. Instead of THIS behavior, it might be worth just appending a
number to the end of the universe directory. This will allow cacheing to
proceed without any problem.