Tested cacheing facility. One caveat.
If cacheing on NFS, decacheing will likely result in a .nfs file being created
in the directory on deletion of the structure and trajectory files. This will
cause deletion of the cache directory to fail. Re-running decache will clear it
up, however.

This is an admissable bug, because there is no point in cacheing on NFS. That
was precisely the point of this feature.

I'm pleased it works robustly.