Now Sim only attaches one universe at a time.
This makes it much easier to use for interactive work. It also makes it
easier to add the selection functionality without making things overly
complicated.

The rationale: even though it useful to be able to associate multiple
universes with a single Sim object, it is not generally useful to be
able to attach more than one at a time. The ability to do so introduces
more complexity in handling the Sim object than is really necessary, and
the payoff is nonexistent. There is no reason why one would need two
verisons of the same trajectory available from the same Sim object at
the same time.

If one really does need to do the above, then one can load another
instance of the same Sim object with a different universe attached. Far
easier to keep track of, and things like AtomGroup selections won't
become confusing ("which universe does this go to...").