Added File prototypes.
We need file syncronization. That is, each time some item that is saved is
modified, it is cleanly written to disk. This must be done at the time of
transaction, and the object must continually refresh its copy of the data in
memory before it makes changes.

The best way to do this is to make classes that provide a clear interface for
these files. This is what the File submodule is for. In fact, the functionality
currently found, for example, in Containers._build_metadata() will be replaced
by these File classes.

Note that these are not meant to be userland interfaces. The user should not
have to interact directly with any File objects. Instead, they interact with
aggregator classes like Core.Info or Core.Data.