Persistent selections now available for Sim.
Now selection strings can be stored as part of the Sim object's metadata. These
selections will be used to build corresponding atomGroups each time a universe
is attached. The selections are re-generated on each re-attach.

The selections dictionary can be as nested as one likes; it is entirely
user-defined. The structure of the dictionary is used to build the dictionary
storing the atomGroup objects. Only dictionaries are allowed; no lists.

Untested yet.