Clean up in preparation for collaboration.
The develop branch will be the main branch for working for both of us.
To make it easier, I will work directly on the Containers, while Sean
can work directly on the Operators. We will both need to put our heads
together for the File workers, as well as the Aggregators for our
respective domains. Utilities can work immediately as grab bag classes.

As for the Database, I'll consider it a container and work on it
directly, as it shouldn't have to directly work with any Operators, but
only with Containers.