Database object now Coordinator.
In the interest of giving a user an interface for accessing any and all
subsets of Containers based on a variety of search and selection
methods, I have created the Coordinator object. This object will "have
a" Database object as currently implemented, which itself is the
interface for all operations on the database stored on disk (via the
DatabaseFile class, which provides syncronized access to the file).

For now I just moved the Database as previously implemented to
Core.Aggregators, but it will need to be cannibalized for methods as
a few might be better fits for Coordinator directly. Also, it used to
inherit from Core.Workers.ObjectCore, which it may actually depend on.
This may also be true for our other Aggregators, so the contents of
ObjectCore will need to be examined in light of the current design.