Hammering away at file locking.
We need operations to be atomic on all files MDS works on, and this requires
reliable file locking. After plenty of digging, the most platform-agnostic but
generally reliable method for locking a file before a read/write operation is
to make a symlink to it. This is because making a symlink also serves as a
check that a symlink is already present, and because making symlinks is an
atomic operation even on NFS mounts (as of version 2).

I am implementing two types of locks: shared locks (*.shlock) and exclusive
locks (*.exlock). A shared lock should be used before a File object reads its
associated file, as it is a signal to other instances that it is okay to read
but not write to that file. An exclusive lock should be used before a File
object writes, as it signals that no read or write operations of other
instances should take place.

Examples:
    There is a call to get the name of a Sim object. This call goes to the
    Metadata Aggregator, and the File object is asked for it. The File object
    applies a shared lock on the file, reads it, then gives back the requested
    name. If a shared lock is already present, it continues anyway...

THERE IS A PROBLEM WITH THIS APPROACH. First, there is no way to atomically
check for either a shared lock or an exclusive lock at the same time. Second,
there is no way to ensure that a shared lock will be in place so long as at
least one reader that requires it is present, since the first to finish or
the original maker of the lock will remove it. It is a problem if instances are
reading the file and the lock is gone.

We are reverting to the simpler but less efficient single-lock mechanism. Only
one instance at a time gets to read or write to a file.

Also, after some experimentation, we will need to open and close file streams
at use. I was hoping we could get away with keeping the filestream for a File
object open using

    filestream = open('filename', 'r+', 1)

which gives line buffering and read/write mode, but though this works great on
a single system, under NFS there is significant delay between when one host
writes and when another host sees the changes. When the filestream is closed,
changes appear immediately. This is probably due to how NFS handles open
filestreams.

Even though doing only one type of lock and having to open and close streams on
each use makes the whole scheme slower, it does make it *simpler*. It will
almost certainly be robust, barring any new potential showstoppers.