Almost finished with File base class.
The File base class contains all functionality required by other File objects
to read and write to their persistent forms on disk with reliable file locking.
It is flexible in that it does not care how the data structure it is
responsible for in memory is translated to disk. This is specified by the
functions self.reader and self.writer, which are to be defined by individual
child classes according to their needs.

This agnosticism is particularly exciting with regard to Data files, since we
should like to make this work for ANY convenient data storage format.