Added prototype aggregators to Coordinator.
Aggregators are interfaces. Container objects "have" aggregators, which provide
most of their functionality. An aggregator changes the state of an Container by
interacting with File objects, since these handle synchronization of the state
of the Container with its representation on disk.

I have a lot of cool ideas for how to make Containers intuitive to use. A Sim
container, for example, will have attributes like the following:

Sim
 -> add
 -> remove
 -> load
 -> universe
 -> selections
 -> tags
 -> categories
 -> data
 -> extra

where many of these attributes are themselves aggregators. For example, add:

add
 -> universe()
 -> tag()
 -> category()
 -> selection()
 -> data()

These functions modify the state of the Sim object accordingly through
interactions with the appropriate File objects. This arrangement is also
intuitive, since it has as the top level the action (verb) followed by the
object of the action (the predicate).

Our other Containers and the Coordinator will be built using a similar schema.