Settled on file names.
Here is the current schema. Container files all have the same state file
(formerly metadata file) name: container.yaml. For now it is still yaml format
for human readability. Log files are all container.log. The reason for using
the same names for all containers is that it makes it easier for the Finder to
locate MDS objects in the filesystem. It need only look for one name, not
three.

Also, we are going to handle container data in the following way. Each time a
data instance is added, a corresponding directory with the name of that
instance will be added to the Container's directory structure. This directory
will contain a file called 'data.yaml', which will include all information
about the stored data instances (also present in this dir). This includes the
reader and writer for these files, timestamps for changes, and other useful
info about the data.