Resorted to magic for ContainerFile.
In order to keep the :meth:`create` call the same for subclasses of File as it is
for File itself, I needed to resort to using kwargs to slurp up all the
elements of the data structure that are given as input. The only other way I
could think of for doing this was to remove the call to :meth:`create`
completely from the File's init method, but this would have left it
fundamentally incomplete.

Neither is an ideally elegant solution, but I think the use of kwargs may
afford us the flexibility we need at this level to get these File classes
finished quickly. After all, these are classes that the user does not even see,
so as long as not too much obfuscating magic is used, this should be fine.