ContainerFile should initialize correctly.
With kwargs magic, everything is in place.

We do need to work out how we will deal with coordinator-less (formerly
database-less) containers. In principle they are fine, and should be built to
work fine without a coordinator. Perhaps if the location of the coordinator is
given as None, it should serve to disable all interactions with a coordinator
that would normally occur?

Perhaps we should make coordinators 'opt-in', in that the a location for one
must be given (even if it doesn't yet exist) in order for a Container to
connect to one. We can make the location of the resulting Container dependent
on this, with an overriding location parameter available as input.