Removed file locking from read and write methods.
Because read and write must be applied in different orders all with one
lock in place for different purposes, I needed to move the locking
mechanism out of the read and write methods themselves. The original
intention was that file locking would automatically occur for these, but
it would take a lot of magic to keep this while addressing our real
needs.

TLDR: Locking must be explicitly done now for File operations.