Added lockit decorator to File class.
The lockit decorator allows one to easily apply a file lock to any
method that requires it without having to write in the same code blocks
for locking and unlocking every time. Since locking and unlocking or
only done at the beginning and at the end, respectively, of any atomic
operation, we need only apply this operator to all setter, getter, and
modder methods.