TestFile class will test multiple files at once.
The File base class requires specification of the reader and writer for
interacting with the version of the data structure on disk. To test the
robustness of this scheme, we will test against as many file types as we
can for our test data structure (consisting of only pure python
objects).

Perhaps its best to eventually make this a list, but for now we will
just access each file object individually in our unit tests.