After getters and setters, refactor create method.
We must be VERY careful of opening multiple instances of a single file.
Even if the handle is reassigned, these old instances are not closed
until garbage collection or the interpreter exits (probably the latter
in most cases). This means that changes may not be flushed before file
locking is released.

TLDR: perform operations as atomically as possible and this will be
avoided.