Removed explicit unlock from decorators.
It shouldn't be necessary to explicitly unlock a file in any use case,
because as soon as the file is closed any lock on it disappears (because
the file handle disappears). In fact, because an unlock would have to
occur before the file is closed, this would be dangerous, since flushes
of writes are not certain until the file is closed.

Don't use unlock without a damn good reason.