Added version for ContainerFile objects.
We need a way to future-proof container objects for when specifications
change. By including a version field, we can include code in new
releases that automatically checks the version field and applies the
necessary transformations to bring the tables up to new standards. This
code can be cascading, too, in the sense that container objects from
many releases back can be transformed to a modern container object
by application of each transformation in sequence.