add_universe built.
As written, won't handle the case where universe already exists.
Probably best to build update_universe as a separate method to avoid
bloating up the adder. We want some kind of update method because it
would be convenient to be able to update topology or trajectory file
paths without having to rebuild the entire definition (including
selections).

Perhaps build individual methods for updating topology and trajectory
files? We can't use these in the add_universe method because we need a
consistent lock applied over the course of adding a universe. Updating
individual topology and trajectory files should be consistent with this
design.