Added default logging to File.
If no logger given to File object (or children), then a default logger
is used with logging level INFO and only the stdout as a stream. This
allows us to write logging output for the File objects with confidence
it will go somewhere. For the purpose of testing, we also want to ensure
File objects can be tested independent of higher level objects.