Added build of selection group to add_universe.
It makes sense that add_universe build all the component data structures
for a universe definitions, including selections. Otherwise,
add_selection must include lots of machinery to check for existence of
things that are completely unnecessary.

Still need to add a try/except for the case where a selection is added
to a nonexistent universe. Better to handle failure cleanly than not at
all.