Added DataFile object.
DataFile is the interface to stored data files, which uses pandas'
HDFStore object as its underlying interface to HDF5. The class itself is
almost trivial in content, since it mostly just maps some methods of
HDFStore onto getters and setters we care about. However, this was
necessary to include proper file locking on all operations we need
for these types of files.

Also, we don't need all the functionality of the HDFStore object. This
class gives us the things we need out of it in a simple interface.