File locking fixed for DataFiles.
Needed to adjust what attribute was used to get file handle.
Also, added data_columns keywords to add and append methods so that all
columns are now indexible.

The where keyword allows for queries on the data without having to read
the whole data structure into memory. This is very cool, but the syntax
is a bit obtuse. We may add a layer later to make this more user
friendly. At the moment it uses exactly what is described in the pandas
docs.