GroupFile now has del_member.
Need to test both this and add_member, as well as other GroupFile
methods. The form of many of these methods was lifted from ContainerFile
methods. Also, can probably speed things up with table.where usage.

Need a method to get a dictionary of a member's stored attributes by
uuid.