Added open file checking to write decorators.
It is convenient to be able to use writing methods within other writing
methods. We must be careful, however, since we must open only one file
descriptor at a time on a given process. We also don't want to change
the lock state of the file in the middle of a method execution.

By adding in the existing machinery for read methods to write methods,
we should be able to freely call write methods within writing methods.
If we try calling write methods from read methods, the fact that the
file is not open for writing will give an error. This is a safety check.