Create methods now perform their own file locking.
Since we now have mechanisms to ensure writers can be called within
writers, added file locking to create methods to keep things proper and
consistent. This is safer, too, since it means that in the (unlikely)
    case that a container is loaded while it is being created a
    partially populated object is not accessed.

Also added initialization of the file handle to ContainerFile.__init__.
We should be able to remove the try/except in the lock decorators
because of this. Need to add similiar initialization to DataFile if I do
this.