Begun work on Aggregators.
Now that files are mostly sorted out, we can move on to Sim and Group
implementation. Sims and Groups each have several types of objects we
call "aggregators", which serve as interfaces to select groupings of
methods or attributes. Development of Container components and their
aggregators will happen in parallel. It is not completely clear cut what
will go where in all cases.