Leaning toward selections being generated on the fly.
It seems to me less confusing to make atom selections regenerated every
time they are accessed. This is because the universe object may change
(e.g. resnames), and these changes should be reflected in our stored
selections. Usually one doesn't change a universe one is doing work on,
but in some cases this may be necessary or preferred. In this case,
stored selections should respect the current state of the universe.