Added create method to Aggregator.
Allows for aggregator-specific initialization without needed to replace
the __init__ method. Fits in with our way of doing things in the File
classes.

Working on Members object, which needs to initialize a list of members.
Members are NOT regenerated on each call, since this is perhaps an
expensive operation. Well, we will see. Toying with the idea of
containers having to find themselves if their state files suddenly go
missing while they are loaded.