Working on Members aggregator.
Members are loaded on initialization of the Members (and by extension,
the Group). The alternative to this is to load members each time member
attributes are required, but this would probably be slow since object
initialization is rather expensive. We can keep this around as an idea,
however.

A problem with loading members first reveals another problem I hadn't
considered until now: what happens when a member's state file is moved
while it is loaded? Well, as of now the member would fail, since it
would not find its file when it needs information from it. A solution is
to implement a decorator for accessors in which a Finder is called to
track down the missing file if it is suddenly missing. Since Finder has
yet to be implemented at all, this will have to wait.