Trying out grouping methods by object.
Instead of using "Add" and "Remove" aggregators, I'm trying out the use
of aggregators that are associated with attributes of a container, like
"Tags" and "Selections". This allows us a bit more flexibility in what
we can include, and helps to keep things more clearly separated.

We'll see if this plays out alright.