Building myriad getters, setters.
Note, we will need a detached keyword for inits on both Sims and Groups.
For Sims it means the default universe is not attached (need to
implement that); for Groups it means members are not loaded. Both are
particularly important so that Groups can load member Sims and Groups
cheaply.