Working on Data aggregator.
This is tricky, because it deals with the filesystem and multiple files
that may or not exist. _get_datafile returns a file handle to the given
file if present. Need to test whether this can cause problems, because
in principle it means we may have as many file handles created as there
are calls to this method. Not sure if there are side-effects.