Working on Data access.
Built decorator for loading and deleting DataFile object on call.
Currently working on remove method. Since we are operating under the
paradigm of one table per file, this needs to delete more than just a
table. It has to delete the file and the directory (if unempty).

Not hard, but I should note that with this way of doing things we don't
even need our table deleter in the DataFile object.