Moved attach/detach to Sim object.
It makes sense to put attach and detach methods at the top level of the
Sim object, since these affect what is mounted at Sim.universe. Methods
that do not affect this mount point but concern Universe definitions
should be placed in Sim.universes. For now this includes the
cache/decache methods (which are still old-style).