Refined Data.remove() so it is lighter weight.
We want to avoid decorating this method with _write_datafile because
this will create a datafile if there isn't one already, which for a
moment would register as a data instance for other instances of this
Container. This method now checks for the existence of a file to handle
before doing anything.

Also, if all rows of a dataset are removed by way of row removals, the
whole dataset is removed in the usual way (directory deleted if nothing
else there).