Considering moving arguments around for Sims and Groups.
Trouble is, name is now required, but it's awkward having it at front.
Regenerating an object now means we now need something for name.

Perhaps we should scrap the idea of allowing Groups and Sims to be
initialized with containers or topologies and trajectories as arguments,
but instead make this available as keywords? This might make things more
flexible but a little more complicated.