Removed caching feature; improved property docs.
Cacheing is a nice idea as an enhancement to the Sim object, but it will
require some thought to implement properly, especially since the Sim
object has to be able to activate and disable a universe based on what
the state file says. Also, the implementation of copying and deletion
was OS dependent, I believe. This feature may make sense to add back in
another day.

Also, added lots of new documentation for properties of Containers.