Not dealing with readthedocs anymore.
Because numexpr refuses to compile on readthedocs using pip, I cannot
get PyTables installed in the virtualenv. The reason I care is because I
want autodoc to grab the docstrings out of my classes, but this can't
happen if the module can't be imported.

For now, added to doc Makefile a way to push the docs to github as the
project page. Can be found at `http://dotsdl.github.io/MDSynthesis/`
for the time being.