Added deactivate method for universes.
Because the existence of a Universe leaves its topology and trajectories
as open files, this can result in "too many open files" errors when
using a large number of Sims and accessing their universes. This method
will hopefully help in those cases where this arises since the garbage
collector can remove the Universe once deactivated.

Hopefully this gets fixed more robustly in MDAnalysis.