Bundle prototype.
We eventually want to have a class that is used to return the results of
queries to a Coordinator or Group. This class would allow no doubles of
any Container, but otherwise it should behave like a list. It can be
used standalone to, for example, take a bunch of paths to statefiles and
make regenerated Sims and Groups out of them.