Added blosc compression (level 5) for pandas data.
This can increase read/write performance if compression succeeds in
decreasing the size of the data written to disk. Will need to experiment
to see if level 5 is generally a good value for most uses.

Note, there is currently not exposed interface for changing the
compression level when storing data.