Moved calls for compression inside pdData.add() and .append().
Putting the compression options into the HDFStore init didn't appear
to make any compression actually happen. Instead, putting these directly
into the calls to HDFStore.put() and HDFStore.append() does appear to
give a compressed file.

In an informal test, the size of a test dataframe on disk decreased by
half under compression with BLOSC level 5, but read time increased by
20%. Perhaps make optimization of the compression level something that
happens during the build process?

In any case, for now we will keep this baked-in compression level using
blosc. Pytables is a requirement of MDS, so this at least will be
present if using MDS at all. Further, an anecdotal reduction of 50%
space usage is to me worth the 20% worse read performance on our system.