Container now a standalone object.
Sim and Group are built on top of the base class Container, which up
until now lacked a proper init to give it a state file. Now it can be
used on its own. Most useful when writing tests, since we need not
separately test tags, categories, and data for Sims and Groups.