Renamed module and submodule to be PEP8 compliant.
Originally we wanted the import statement to mirror that of MDAnalysis::

    import MDAnalysis
    import MDSynthesis

But in the interst of being PEP8 compliant, the package will now import
as::

    import mdsynthesis

All submodules have been made lowercase as well. The overall repository
name on github will keep the capitalization, and we'll keep the
capitalization in the docs, too. In other words, we will distinguish
between the package name and the module name.