Lots of work on making uuid-including state files work.
Including uuids in state file names instead of inside the file should
turn finding containers in a filesystem into a path parsing task instead
of an expensive file-reading task. However, in the meantime it means
that methods for loading state files have to use globbing. It also
introduces the tricky (though perhaps useful) possibility of more than
one state file occupying a directory. It's probably best this be
avoided, but I am implementing these changes to accommodate that
scenario gracefully. It *should* work.